Refuse to compare types with type families
'caseElemNonReachable' considers a branch unreachable when a branch
coerces a type to another. For example, a branch coercing `Bool ~ Int`
would be considered unreachable. It would also consider `F Char ~ Int` to
be unreachable, even if `F` was a type family eventually resolving to
`Int`.

Regrettably, I haven't been able to find a small test case.