Revert adding 'reduceNonRepPrim' to 'flatten'
The historical record is lacking, but this was probably added as a
poor man's solution to cases where register was expecting a constant
argument but wasn't getting any. After introducing a proper solution
(see the work on constantSpec), we hopefully don't need this anymore.

Reverted due to it causing problems in an external codebase:
'flattenLet' expects a given expression to be in NF, but
reduceNonRepPrim might violate this invariant.