Update Dockerfile and CI scripts
The Dockerfile is now based on Ubuntu 18.04 LTS, and allows testing
GHC versions up to and including 8.10.1. As an added bonus the
resulting image is also smaller, as the Dockerfile now uses multiple
build stages.

The scripts now expect the $GHC and $CABAL variables to be set, and
calls these instead of `ghc` or `cabal` directly. This is to ensure
that each job in the CI configuration script is explicit about the
versions of these programs they run with (and that this information
is easy to find).